A download parameter instead of with_download
This was inspired by GH-43. It removes the with_download method and
instead introduces a download parameter in the embedder interface.

If download=True, the shared constructor downloads the required files,
otherwise it checks the user provided them as parameters (and reminds
the user if they didn't).

This doesn't touch the pipeline's own download logic.

Nice side effect: 110 insertions(+), 134 deletions(-) (including new tests)