Add json schema for pipeline
This isn't usable for our target audience as it currently is, but might still be good to have in git.

One thing we could try is running the python json schema validator over the provided pipeline to give helpful error messages.

I've added the file besides the parameter blueprints because it's essentially a machine readable way to write that.