Use humanize on file sizes
humanize makes nice output and has zero dependencies

Old:

```
INFO The minimum expected size for the reduced_embedding_file is 0.010MB.
INFO The minimum expected size for the embedding_file is 0.077MB.
INFO You are going to generate a total of 0.087MB of embeddings, and have 430904.967MB available at test_1.
```

New:

```
INFO The minimum expected size for the reduced_embedding_file is 10.2 kB.
INFO The minimum expected size for the embedding_file is 76.8 kB.
INFO You are going to generate a total of 87.0 kB of embeddings, and have 430.9 GB available at test_1.
```