Store prediction results in temporary directory
Will maybe be changed in the future. However, we need some dedicated
directory for storing the results anyway.

Probably found the issue for the connection loss to the broker: rabbitmq
logs a loss of connection because of a missing heartbeat message. Didn't
find a way to fix that yet, this still needs to be done